+++ src/core/include/mp-units/framework/construction_helpers.h
@@ -22,26 +22,46 @@
 
 #pragma once
 
-// IWYU pragma: begin_exports
-#include <mp-units/framework/compare.h>
-#include <mp-units/framework/construction_helpers.h>
-#include <mp-units/framework/customization_points.h>
-#include <mp-units/framework/dimension.h>
-#include <mp-units/framework/dimension_concepts.h>
-#include <mp-units/framework/expression_template.h>
-#include <mp-units/framework/magnitude.h>
 #include <mp-units/framework/quantity.h>
-#include <mp-units/framework/quantity_cast.h>
-#include <mp-units/framework/quantity_concepts.h>
 #include <mp-units/framework/quantity_point.h>
-#include <mp-units/framework/quantity_point_concepts.h>
-#include <mp-units/framework/quantity_spec.h>
-#include <mp-units/framework/quantity_spec_concepts.h>
-#include <mp-units/framework/reference.h>
+#include <mp-units/framework/reference_concepts.h>
 #include <mp-units/framework/representation_concepts.h>
-#include <mp-units/framework/symbol_text.h>
-#include <mp-units/framework/system_reference.h>
-#include <mp-units/framework/unit.h>
-#include <mp-units/framework/unit_concepts.h>
-#include <mp-units/framework/value_cast.h>
-// IWYU pragma: end_exports
+
+#ifndef MP_UNITS_IN_MODULE_INTERFACE
+#include <type_traits>
+#endif
+
+namespace mp_units {
+
+template<Reference R>
+struct delta_ {
+  template<typename Rep>
+    requires RepresentationOf<std::remove_cvref_t<Rep>, get_quantity_spec(R{}).character>
+  [[nodiscard]] constexpr quantity<R{}, std::remove_cvref_t<Rep>> operator()(Rep&& lhs) const
+  {
+    return quantity{std::forward<Rep>(lhs), R{}};
+  }
+};
+
+template<Reference R>
+struct absolute_ {
+  template<typename Rep>
+    requires RepresentationOf<std::remove_cvref_t<Rep>, get_quantity_spec(R{}).character>
+  [[nodiscard]] constexpr quantity_point<R{}, default_point_origin(R{}), std::remove_cvref_t<Rep>> operator()(
+    Rep&& lhs) const
+  {
+    return quantity_point{quantity{std::forward<Rep>(lhs), R{}}};
+  }
+};
+
+MP_UNITS_EXPORT_BEGIN
+
+template<Reference auto R>
+inline constexpr delta_<MP_UNITS_REMOVE_CONST(decltype(R))> delta{};
+
+template<Reference auto R>
+inline constexpr absolute_<MP_UNITS_REMOVE_CONST(decltype(R))> absolute{};
+
+MP_UNITS_EXPORT_END
+
+}  // namespace mp_units