+++ src/core/include/mp-units/framework/quantity.h
@@ -96,13 +96,8 @@ template<typename Func, Quantity Q1, Quantity Q2>
 using common_quantity_for = quantity<common_reference(Q1::reference, Q2::reference),
                                      std::invoke_result_t<Func, typename Q1::rep, typename Q2::rep>>;
 
-template<auto T, auto R>
-concept SameOriginalReferenceAs =
-  DeltaReference<MP_UNITS_REMOVE_CONST(decltype(T))> && Reference<MP_UNITS_REMOVE_CONST(decltype(R))> &&
-  detail::SameReference<remove_reference_specifier(T), R>;
-
 template<auto R1, auto R2, typename Rep1, typename Rep2>
-concept SameValueAs = detail::SameOriginalReferenceAs<R1, R2> && std::same_as<Rep1, Rep2>;
+concept SameValueAs = detail::SameReference<R1, R2> && std::same_as<Rep1, Rep2>;
 
 }  // namespace detail
 
@@ -133,25 +128,25 @@ public:
   [[nodiscard]] static constexpr quantity zero() noexcept
     requires requires { quantity_values<rep>::zero(); }
   {
-    return {quantity_values<rep>::zero(), detail::make_delta(R)};
+    return {quantity_values<rep>::zero(), R};
   }
 
   [[nodiscard]] static constexpr quantity one() noexcept
     requires requires { quantity_values<rep>::one(); }
   {
-    return {quantity_values<rep>::one(), detail::make_delta(R)};
+    return {quantity_values<rep>::one(), R};
   }
 
   [[nodiscard]] static constexpr quantity min() noexcept
     requires requires { quantity_values<rep>::min(); }
   {
-    return {quantity_values<rep>::min(), detail::make_delta(R)};
+    return {quantity_values<rep>::min(), R};
   }
 
   [[nodiscard]] static constexpr quantity max() noexcept
     requires requires { quantity_values<rep>::max(); }
   {
-    return {quantity_values<rep>::max(), detail::make_delta(R)};
+    return {quantity_values<rep>::max(), R};
   }
 
   // construction, assignment, destruction
@@ -160,7 +155,7 @@ public:
   quantity(quantity&&) = default;
   ~quantity() = default;
 
-  template<typename Value, DeltaReference R2>
+  template<typename Value, Reference R2>
     requires detail::SameValueAs<R2{}, R, std::remove_cvref_t<Value>, Rep>
   constexpr quantity(Value&& v, R2) : numerical_value_is_an_implementation_detail_(std::forward<Value>(v))
   {
@@ -171,24 +166,6 @@ public:
             detail::QuantityConvertibleTo<quantity<R2{}, std::remove_cvref_t<Value>>, quantity>
   constexpr quantity(Value&& v, R2) : quantity(quantity<R2{}, std::remove_cvref_t<Value>>{std::forward<Value>(v), R2{}})
   {
-  }
-
-  template<typename Value, DeltaReference R2>
-    requires(!Reference<R2>) && (!detail::SameValueAs<R2{}, R, std::remove_cvref_t<Value>, Rep>) &&
-            detail::QuantityConvertibleTo<
-              quantity<detail::remove_reference_specifier(R2{}), std::remove_cvref_t<Value>>, quantity>
-  constexpr quantity(Value&& v, R2) :
-      quantity(
-        quantity<detail::remove_reference_specifier(R2{}), std::remove_cvref_t<Value>>{std::forward<Value>(v), R2{}})
-  {
-  }
-
-  template<typename Value, Reference R2>
-    requires(!DeltaReference<R2>)
-  constexpr quantity(Value&&, R2)
-  {
-    static_assert(DeltaReference<R2>,
-                  "References using offset units (e.g., temperatures) must be explicitly qualified with `delta`");
   }
 
   template<detail::QuantityConvertibleTo<quantity> Q>
@@ -206,8 +183,8 @@ public:
                                           convert_explicitly> ||
                      // NOLINTNEXTLINE(google-explicit-constructor, hicpp-explicit-conversions)
                      !std::convertible_to<typename quantity_like_traits<Q>::rep, Rep>) quantity(const Q& q) :
-      quantity(::mp_units::quantity{quantity_like_traits<Q>::to_numerical_value(q).value,
-                                    detail::make_delta(quantity_like_traits<Q>::reference)})
+      quantity(
+        ::mp_units::quantity{quantity_like_traits<Q>::to_numerical_value(q).value, quantity_like_traits<Q>::reference})
   {
   }
 
@@ -306,7 +283,7 @@ public:
       } -> std::common_with<rep>;
     }
   {
-    return ::mp_units::quantity{+numerical_value_is_an_implementation_detail_, detail::make_delta(reference)};
+    return ::mp_units::quantity{+numerical_value_is_an_implementation_detail_, reference};
   }
 
   [[nodiscard]] constexpr QuantityOf<quantity_spec> auto operator-() const
@@ -316,7 +293,7 @@ public:
       } -> std::common_with<rep>;
     }
   {
-    return ::mp_units::quantity{-numerical_value_is_an_implementation_detail_, detail::make_delta(reference)};
+    return ::mp_units::quantity{-numerical_value_is_an_implementation_detail_, reference};
   }
 
   template<typename Q>
@@ -338,7 +315,7 @@ public:
       } -> std::common_with<rep>;
     }
   {
-    return ::mp_units::quantity{numerical_value_is_an_implementation_detail_++, detail::make_delta(reference)};
+    return ::mp_units::quantity{numerical_value_is_an_implementation_detail_++, reference};
   }
 
   template<typename Q>
@@ -360,7 +337,7 @@ public:
       } -> std::common_with<rep>;
     }
   {
-    return ::mp_units::quantity{numerical_value_is_an_implementation_detail_--, detail::make_delta(reference)};
+    return ::mp_units::quantity{numerical_value_is_an_implementation_detail_--, reference};
   }
 
   // compound assignment operators
@@ -463,16 +440,6 @@ template<typename Value, Reference R>
   requires RepresentationOf<Value, get_quantity_spec(R{}).character>
 quantity(Value v, R) -> quantity<R{}, Value>;
 
-template<typename Value, DeltaReference R>
-  requires(!Reference<R>) &&
-          RepresentationOf<Value, get_quantity_spec(detail::remove_reference_specifier(R{})).character>
-quantity(Value v, R) -> quantity<detail::remove_reference_specifier(R{}), Value>;
-
-// the below is needed only to fire static_asserts in the constructor
-template<typename Value, Reference R>
-  requires(!DeltaReference<R>) && RepresentationOf<Value, get_quantity_spec(R{}).character>
-quantity(Value v, R) -> quantity<R{}, Value>;
-
 template<QuantityLike Q>
 explicit(
   is_specialization_of<decltype(quantity_like_traits<Q>::to_numerical_value(std::declval<Q>())), convert_explicitly>)
@@ -487,7 +454,7 @@ template<auto R1, typename Rep1, auto R2, typename Rep2>
   const ret ret_lhs(lhs);
   const ret ret_rhs(rhs);
   return quantity{ret_lhs.numerical_value_ref_in(ret::unit) + ret_rhs.numerical_value_ref_in(ret::unit),
-                  detail::make_delta(ret::reference)};
+                  ret::reference};
 }
 
 template<auto R1, typename Rep1, auto R2, typename Rep2>
@@ -498,7 +465,7 @@ template<auto R1, typename Rep1, auto R2, typename Rep2>
   const ret ret_lhs(lhs);
   const ret ret_rhs(rhs);
   return quantity{ret_lhs.numerical_value_ref_in(ret::unit) - ret_rhs.numerical_value_ref_in(ret::unit),
-                  detail::make_delta(ret::reference)};
+                  ret::reference};
 }
 
 template<auto R1, typename Rep1, auto R2, typename Rep2>
@@ -511,15 +478,14 @@ template<auto R1, typename Rep1, auto R2, typename Rep2>
   const ret ret_lhs(lhs);
   const ret ret_rhs(rhs);
   return quantity{ret_lhs.numerical_value_ref_in(ret::unit) % ret_rhs.numerical_value_ref_in(ret::unit),
-                  detail::make_delta(ret::reference)};
+                  ret::reference};
 }
 
 template<auto R1, typename Rep1, auto R2, typename Rep2>
   requires detail::InvocableQuantities<std::multiplies<>, quantity<R1, Rep1>, quantity<R2, Rep2>>
 [[nodiscard]] constexpr Quantity auto operator*(const quantity<R1, Rep1>& lhs, const quantity<R2, Rep2>& rhs)
 {
-  return quantity{lhs.numerical_value_ref_in(get_unit(R1)) * rhs.numerical_value_ref_in(get_unit(R2)),
-                  detail::make_delta(R1 * R2)};
+  return quantity{lhs.numerical_value_ref_in(get_unit(R1)) * rhs.numerical_value_ref_in(get_unit(R2)), R1 * R2};
 }
 
 template<auto R, typename Rep, typename Value>
@@ -527,7 +493,7 @@ template<auto R, typename Rep, typename Value>
           detail::InvokeResultOf<get_quantity_spec(R).character, std::multiplies<>, Rep, const Value&>
 [[nodiscard]] constexpr QuantityOf<get_quantity_spec(R)> auto operator*(const quantity<R, Rep>& q, const Value& v)
 {
-  return quantity{q.numerical_value_ref_in(get_unit(R)) * v, detail::make_delta(R)};
+  return quantity{q.numerical_value_ref_in(get_unit(R)) * v, R};
 }
 
 template<typename Value, auto R, typename Rep>
@@ -535,7 +501,7 @@ template<typename Value, auto R, typename Rep>
           detail::InvokeResultOf<get_quantity_spec(R).character, std::multiplies<>, const Value&, Rep>
 [[nodiscard]] constexpr QuantityOf<get_quantity_spec(R)> auto operator*(const Value& v, const quantity<R, Rep>& q)
 {
-  return quantity{v * q.numerical_value_ref_in(get_unit(R)), detail::make_delta(R)};
+  return quantity{v * q.numerical_value_ref_in(get_unit(R)), R};
 }
 
 template<auto R1, typename Rep1, auto R2, typename Rep2>
@@ -543,8 +509,7 @@ template<auto R1, typename Rep1, auto R2, typename Rep2>
 [[nodiscard]] constexpr Quantity auto operator/(const quantity<R1, Rep1>& lhs, const quantity<R2, Rep2>& rhs)
 {
   MP_UNITS_EXPECTS_DEBUG(rhs != rhs.zero());
-  return quantity{lhs.numerical_value_ref_in(get_unit(R1)) / rhs.numerical_value_ref_in(get_unit(R2)),
-                  detail::make_delta(R1 / R2)};
+  return quantity{lhs.numerical_value_ref_in(get_unit(R1)) / rhs.numerical_value_ref_in(get_unit(R2)), R1 / R2};
 }
 
 template<auto R, typename Rep, typename Value>
@@ -553,7 +518,7 @@ template<auto R, typename Rep, typename Value>
 [[nodiscard]] constexpr QuantityOf<get_quantity_spec(R)> auto operator/(const quantity<R, Rep>& q, const Value& v)
 {
   MP_UNITS_EXPECTS_DEBUG(v != quantity_values<Value>::zero());
-  return quantity{q.numerical_value_ref_in(get_unit(R)) / v, detail::make_delta(R)};
+  return quantity{q.numerical_value_ref_in(get_unit(R)) / v, R};
 }
 
 template<typename Value, auto R, typename Rep>
@@ -562,7 +527,7 @@ template<typename Value, auto R, typename Rep>
 [[nodiscard]] constexpr QuantityOf<inverse(get_quantity_spec(R))> auto operator/(const Value& v,
                                                                                  const quantity<R, Rep>& q)
 {
-  return quantity{v / q.numerical_value_ref_in(get_unit(R)), detail::make_delta(::mp_units::one / R)};
+  return quantity{v / q.numerical_value_ref_in(get_unit(R)), ::mp_units::one / R};
 }
 
 template<auto R1, typename Rep1, auto R2, typename Rep2>