+++ src/core/include/mp-units/framework/unit_concepts.h
@@ -210,4 +210,11 @@ concept UnitCompatibleWith =
   Unit<U> && Unit<MP_UNITS_REMOVE_CONST(decltype(FromU))> && QuantitySpec<MP_UNITS_REMOVE_CONST(decltype(QS))> &&
   (!AssociatedUnit<U> || UnitOf<U, QS>)&&detail::UnitConvertibleTo<FromU, U{}>;
 
+namespace detail {
+
+template<typename T>
+concept OffsetUnit = Unit<T> && requires { T::point_origin; };
+
+}
+
 }  // namespace mp_units